R2-2017
* Made audio button visible

Changed the application.scss file to show the audio collapse button. Before it was just plain white and couldn't be seen.

* Displays all buttons (audio)

Shows all audio that is assigned to a stage under the audio tab as buttons.

* Image Tag fix

Image tag fixed as per request

* Added audio ID

added audio id variable, for access via coffee file

* Changed maning of audio ID

same as above

* Coffee file (not complete)

The coffee file that has trouble reading the audio type.

* Coffee file (Incomplete)

last upload was wrong file.

* Test Database Config

Change the test database to behave more like production

* Fix table background

This rule doesn't seem to apply to anything besides the things it shouldn't

* Fix Config

* Redis is now used for test and production
* sqlite3 is only used in development

* Issue #220: Change Panel Title Styling

* Panel titles are now white
* Panel title links are now black on hover

* wav file working

an update where .wav files work, and all users can hear it

* Removed audio collapse

removed an unnecessary audio collapse, all audio options are now displayed once the audio tab is clicked on

* All users listen to audio

All users can listen to the audio that is played

* Bug fix

fixed a bug where the audio file would play multiple times.

* Sortable Columns

Sorts the rows by clicking the column header names

* Changed the table row color

Previously the header is upstageGreen and the background color is upstageGreen, wasnt able to see the column names on it hence for now made it black.

* Review update

removal of minor white space

* Review update

white space removal

* Added comments

* Moved the js function to a file

Sorttable.js file

* Removed inline script

Incomplete file, Still having problems with reading the function with <% javascript_include_tag "sorttable.js" %>, this will throw this error
Rails.application.config.assets.precompile += %w( sorttable.js ) need this code to be entered in config/initializers/assets.rb but still doesnt call the function after that was entered.

* Reverted the panel-title to white

* Precompile sorttable.js

Rails.application.config.assets.precompile += %w( sorttable.js )

* Changed the syntax for javascript_include_tag

Using <%= ... > instead of <% ... >

* Removed play all

Play all button is removed- no extra buttons

* comment code and play all removal

all comment code asked to be taken out is gone, and the play all button functionality is now gone too

* Added TypeError Code

Still not able to sort after using if (typeof rows !== 'null'){...}

* Created Mirror and Button div

Added div segments to distinguish between mirror and button list in the avatar tab

* Granted file access for mirror

allowed the mirror under the 'hold' function to access the picture that will be displayed inside.

* Mirror border and alignment

added the scss file that will center the avatar image inside the mirror
Added a border around the mirror in UpStage green

* Added the mirror function

Added the rest of the mirror function that;
Shows the image of the avatar in the mirror when an avatar is being held,
Removes the image when the avatar is dropped,
Switches image depending on the avatar being held

* Assigned the table variable 

Table assigned as first instruction in the function rather than outside of the function.

* Changed color to white

* Updated audio play (Not Working)

updated audio files that don't use ajax, and with no extra code in the theater.coffee file
Not Currently working - there is a bug where "audio_id" data is not retrieved from the audio_channel

* Fix Review Issue.

Javascript is a case-sensitive language.  Therefore, audioID and
audioId are not the same thing.

* Final audio update

fixed indentation
removed logging (debugging code)

* update drop down menu on adding media

* Can upload an image as a backdrop and assign tags

* Modified files with syntax errors

* fix spelling

* Updated User ID

switched to "current_user.id" instead of "current_user.nickname" to prevent problems in the future as per Paul's request.

* Rspecs for models: AvatarStage, BackdropTag, StageBackdrop, StageSound, UserRole

* Revert "Rspecs for models: AvatarStage, BackdropTag, StageBackdrop, StageSound, UserRole"

This reverts commit 800e81b8a3bdec388348a1203ea409b714d07cde.

* Unit tests for issue212, last two tests in spec/models/backdrop_tag_spec.rb will fail

* fixed uniqueness validations

* backdrop model Rspec unit tests

* Reorganize Configuration

It makes sense to only rely on MySQL for production environments.

Also, it's better to try and declare groups once for simple code.

* Deprecations Warning

Removing the method ActiveSupport.halt_callback_chains_on_return_false resolved the warning. Not sure if it is the right approach?

* Removed controls attribute

* Create .travis.yml

* Add ruby versio

Set to 2.4.1

* Add CI status to README

* Issue#213 Assigning backdrop to stage

* sorted assign backdrop to stage list

* sorted backdrop list when adding to stage

when backdrops are added to the stage it will be sorted now

* Removed unnecessary code

* removed css and update bootstrap

added some bootstrap classes allowing different size window viewing

* sound bug fixed

* avatars alignment fixed for different size screens

* Bump Ruby Version

* Update README.md

Bump Ruby version

* avatar list ordered

* Password Reset Form Changes

- Disables browser autocomplete
- Makes it clearer that a new password is required

* changed avatar code

keep the code same to Sound and Backdrop, Removed some code that is not used. Fixed Stage tag not assigning to the stage

* Update README.md

Point the Travis tag at master in anticipation of release