Issue #214 (#252)
* added Backdrop Carousel

* Changing backdrop on stage

Still buggy. When getting backdrop from the database it doesnt return anything. Clicking on any picture will only give you id 0

* update

Removed the allocations. Backdrop_id was not passing in correctly, now able to get the id

* draw backdrop onto canvas

using image to hold backdrop, manipulate the array to show or hide the backdrop when being clicked

* issue with backdrop showing only 1 upon loading first time is fixed

* fixed the issue when clicking on backdrops

clicking on the new backdrop once will draw the image to the canvas

* removed debugging code

* Replaced JavaScript with coffee script and

* img_tag is not using thumb. Also removed the @ in the coffeeScript