R5-2018 (#320)
* Bump Ruby version to 2.5.0 (#311)

* Issue #312 (#313)

CVE-2018-8048

* Issue #315 (#316)

According to https://github.com/rails/rails/issues/31517 updating the nio4r gem is the solution.  It appears to work.

* Issue #210 (#310)

* Table for media items with the respective shown on the tag page

* tags added into table, should be media items

* Table show, unable to get media showing in table

* Table should display backdrop, sound and stage media. Avatars not showing yet

* avatar should display in table. Table does not sort yet

* Table now sorts by name or type when clicking the column heading

* indentation fixed to show.html.erb and tags.all in show method in tags controller removed

* Removed white spacing and adjust indentation

* Remove Spaces

In HTML, there shouldn't be spaces either side of an ```=``` like you would see in Ruby or Javascript

* Update Rails 5.1.6

Minor release. Applies latest security and bug fixes

* Issue #254 (#317)

* Changed delete messages. Delete messages are now
more consistent.

* Inconsistent messages are now of the same style as other messages.

* Wording changed in delete message. "from the server"
removed and changed to "immediately"

* Fix Typo

* Issue #205 (#318)

* Changed the name of edit button to save

* Remove active class in avatar and backdrop tab

* Chat name adjusted
If the player is holding an avatar and they message on the chat, their name will be displayed as their avatar nickname
If the player is not holding an avatar then their name displayed on the chat will the their username

* Chat name adjusted
If the player is holding an avatar and they message on the chat, their name will be displayed as their avatar nickname
If the player is not holding an avatar then their name displayed on the chat will the their username

* Remove log message

* Changed variable name

* Avatar and backdrop now shows image when you click the tab

* Added new migration file