Enforce Lowercase Usernames
Because usernames are now being used for logins, we must be mindful of how we match.  An ActiveRecord query like find_by_username will assume case sensitivity and thus won't match 'User' and 'user'.

Solution: ensure all existing usernames are stored in lowercase (and going forward) and also force the downcasing of the uid param